IPv6 Support (#127)

+++ cmd/internal/vlan/reservation.go
@@ -17,13 +17,13 @@ func (m Mapping) ReserveVlanIDs(n uint16) ([]uint16, error) {
 
 func (m Mapping) reserveVlanIDs(min, max, n uint16) ([]uint16, error) {
 	maxVlans := max - min + 1
-	if uint16(len(m))+n > maxVlans {
+	if len(m)+int(n) > int(maxVlans) {
 		return nil, fmt.Errorf("can not reserve %d vlan ids, %d are already taken and %d possible at max", n, len(m), maxVlans)
 	}
 	// scan vlan id range for n free vlan ids
 	r := []uint16{}
 	for i := min; i <= max; i++ {
-		if uint16(len(r)) >= n {
+		if len(r) >= int(n) {
 			break
 		}
 		if _, has := m[i]; !has {
